Move blur_array to feature/src/common.
